add docs to src

+++ src/dispeRse_init.c
@@ -1,7 +1,9 @@
+/* Registers the function run_model exported to R */
+
 #include <stdlib.h>
 #include <R_ext/Rdynload.h>
 
-extern void run_model(int*, int*, double*, double*, int*, double*, double*, int*, int*, int*, int*, int*, double*, int*, int*, double*);
+extern void run_model(int*, int*, double*, int*, double*, int*, int*, int*, int*, int*, int*, double*, double*, double*, int*, double*);
 
 static const R_CMethodDef CEntries[] = {
     {"run_model", (DL_FUNC) &run_model, 16},
